Move to TileExportService
